+++ CHANGELOG.md
@@ -2,6 +2,9 @@
 
 ## Unreleased
 
+* Drop support for rubies under 2.3.
+  PR [#1070](https://github.com/activerecord-hackery/ransack/pull/1070)
+
 ## 2.3.0 - 2019-08-18
 
 * Arabic translations PR [979](https://github.com/activerecord-hackery/ransack/pull/979)