+++ CHANGELOG.md
@@ -1,12 +1,14 @@
 # Change Log
 
-## Unreleased
+## 2.3.1 - 2010-01-01
 
 * Drop support for Active Record 5.0, 5.1, and 5.2.0.
   PR [#1073](https://github.com/activerecord-hackery/ransack/pull/1073)
-  
+
 * Drop support for rubies under 2.3.
   PR [#1070](https://github.com/activerecord-hackery/ransack/pull/1070)
+
+  ... and others
 
 ## 2.3.0 - 2019-08-18